requirements.txt
nose
django-nose
-e git://github.com/Almad/django-sane-testing.git#egg=djangosanetesting